+++ src/index.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-import { Command } from "commander";
+import { Command, Option } from "commander";
 
 import { logs, restart, shutdown, start, status } from "./commands";
 import { startDaemon } from "./commands/start-daemon.command";
@@ -18,9 +18,9 @@ program.command("logs [name...]").action((names) => {
 
 program
     .command("status [name...]")
-    .option("--refresh", "Keep status open and refresh periodically")
+    .addOption(new Option("-i, --interval [seconds]", "Keep status open and refresh periodically at given interval").preset("1"))
     .action((names, options) => {
-        status({ names, refresh: !!options.refresh });
+        status({ names, interval: options.interval ? parseInt(options.interval) : undefined });
     });
 
 program.command("restart [name...]").action((names) => {